third_party/blink/web_tests/fast/events/page-visibility-transition-test-expected.txt

This test checks that Page Visibility state values are correct and the event changes are fired correctly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.visibilityState is "visible"
PASS document.hidden is false
PASS document.visibilityState is "hidden"
PASS document.hidden is true
PASS document.visibilityState is "hidden"
PASS document.hidden is true
PASS document.visibilityState is "visible"
PASS document.hidden is false
PASS successfullyParsed is true

TEST COMPLETE
